fix: refactoring
